pyproject.toml
[project]
name = "rate-design-platform"
version = "0.0.0"
description = "Simulation platform for electrical rate design."
authors = [{ name = "Switchbox", email = "hello@switch.box" }]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "buildstock-fetch>=1.6.1",
    "cairo",
    "python-dotenv>=1.0.0",
    "cloudpathlib[s3]>=0.23.0",
    "geopandas>=1.0.0",
    "numpy",
    "pandas",
    "polars",
    "prek>=0.3.2",
    "pyarrow",
    "pygris>=0.2.1",
    "s3fs>=2026.2.0",
    "cloudpathlib>=0.23.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "sodapy>=2.2.0",
    "questionary>=2.1.1",
    "rich>=14.3.2",
    "tqdm>=4.67.3",
    "python-calamine>=0.6.1",
    "fastexcel>=0.19.0",
]

[project.urls]
Repository = "https://github.com/switchbox-data/rate-design-platform"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "ruff>=0.11.5",
    "ty>=0.0.1a21",
    "deptry>=0.20.0",
    "ipykernel>=7.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["data*", "rate_design*", "utils*"]

[tool.setuptools.package-data]
"rate_design.ny.hp_rates" = [
    "scenarios/*.yaml",
    "config/tariffs/electricity/*.json",
    "config/tariffs/gas/*.json",
    "config/tariff_maps/electric/*.csv",
    "config/tariff_maps/gas/*.csv",
]
"rate_design.ri.hp_rates" = [
    "config/tariffs/electricity/*.json",
    "config/tariffs/gas/*.json",
    "config/tariff_maps/electric/*.csv",
    "config/tariff_maps/gas/*.csv",
]
"utils.post" = ["data/*.yaml"]

[tool.uv.sources]
cairo = { git = "https://github.com/NatLabRockies/CAIRO", rev = "9b52c91" }
